+++ resources/views/bootstrap-5/includes/search.blade.php
@@ -8,13 +8,11 @@
         >
 
         @if (isset($filters['search']) && strlen($filters['search']))
-            <div class="input-group-append">
-                <button wire:click="$set('filters.search', null)" class="btn btn-outline-secondary" type="button">
-                    <svg style="width:.75em;height:.75em" xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" stroke="currentColor">
-                        <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M6 18L18 6M6 6l12 12" />
-                    </svg>
-                </button>
-            </div>
+            <button wire:click="$set('filters.search', null)" class="btn btn-outline-secondary" type="button">
+                <svg style="width:.75em;height:.75em" xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" stroke="currentColor">
+                    <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M6 18L18 6M6 6l12 12" />
+                </svg>
+            </button>
         @endif
     </div>
 @endif