+++ resources/views/bootstrap-5/includes/sorting-pills.blade.php
@@ -14,7 +14,7 @@
                     wire:click.prevent="removeSort('{{ $col }}')"
                     class="text-white ms-2"
                 >
-                    <span class="sr-only">@lang('Remove sort option')</span>
+                    <span class="visually-hidden">@lang('Remove sort option')</span>
                     <svg style="width:.5em;height:.5em" stroke="currentColor" fill="none" viewBox="0 0 8 8">
                         <path stroke-linecap="round" stroke-width="1.5" d="M1 1l6 6m0-6L1 7" />
                     </svg>
@@ -25,7 +25,7 @@
         <a
             href="#"
             wire:click.prevent="resetSorts"
-            class="badge rounded-pill bg-light"
+            class="badge rounded-pill bg-light text-dark text-decoration-none"
         >
             @lang('Clear')
         </a>