+++ resources/views/tailwind/includes/table.blade.php
@@ -29,7 +29,7 @@
 
     <x-slot name="body">
         @if (count($bulkActions) && $selectPage && $rows->total() > $rows->count())
-            <x-livewire-tables::table.row wire:key="row-message" class="bg-primary-50">
+            <x-livewire-tables::table.row wire:key="row-message" class="bg-indigo-50">
                 <x-livewire-tables::table.cell :colspan="count($bulkActions) ? count($columns) + 1 : count($columns)">
                     @unless ($selectAll)
                         <div>